requirements.txt
torch
torchvision
matplotlib
pillow
sklearn